#25 拆出语法分析器
